Moved Reflectionhelper to SIL.Core so it doesn't depend on System.Windows.Forms (and it makes more sense there).
