+++ SIL.Scripture.Tests/SIL.Scripture.Tests.csproj
@@ -341,10 +341,6 @@
       <Project>{F71BA7B9-D9DC-4F8C-A307-87B503D0E05B}</Project>
       <Name>SIL.Scripture</Name>
     </ProjectReference>
-    <ProjectReference Include="..\SIL.Windows.Forms\SIL.Windows.Forms.csproj">
-      <Project>{db44f49c-d8c6-434f-81ed-28ea5c9e8195}</Project>
-      <Name>SIL.Windows.Forms</Name>
-    </ProjectReference>
   </ItemGroup>
   <ItemGroup>
     <None Include="Resources\eng.vrs" />
