+++ SIL.Windows.Forms/AppColorPalette/AppColorPaletteExtender.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.Drawing;
 using System.Windows.Forms;
+using SIL.Reflection;
 
 namespace SIL.Windows.Forms.AppColorPalette
 {
