+++ SIL.Windows.Forms/SIL.Windows.Forms.csproj
@@ -46,7 +46,8 @@
     <WarningLevel>4</WarningLevel>
     <DocumentationFile>..\output\Debug\SIL.Windows.Forms.xml</DocumentationFile>
     <PlatformTarget>AnyCPU</PlatformTarget>
-    <TargetFrameworkProfile></TargetFrameworkProfile>
+    <TargetFrameworkProfile>
+    </TargetFrameworkProfile>
     <NoWarn>1591</NoWarn>
     <Prefer32Bit>false</Prefer32Bit>
   </PropertyGroup>
@@ -637,7 +638,6 @@
     </Compile>
     <Compile Include="PropertyBag-NoTests\FilePathEditor.cs" />
     <Compile Include="PropertyBag-NoTests\PropertyBag.cs" />
-    <Compile Include="ReflectionHelper.cs" />
     <Compile Include="Registration\Registration.Designer.cs">
       <AutoGen>True</AutoGen>
       <DesignTimeSharedInput>True</DesignTimeSharedInput>
